Final_week/requirements.txt
numpy
tensorflow
gdown
Pillow
streamlit
opencv-python